IntelliNews/requirements.txt
streamlit
python-dotenv
azure-ai-textanalytics
azure-ai-translation-text
azure-cognitiveservices-speech